udiskslinuxfilesystem: Refactor mount options computation out
... into a separate file.
